common: move pxCheckBox to the main project
