+++ common/common.vcxproj
@@ -54,7 +54,6 @@
     <ClCompile Include="PrecompiledHeader.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="pxCheckBox.cpp" />
     <ClCompile Include="pxRadioPanel.cpp" />
     <ClCompile Include="pxStaticText.cpp" />
     <ClCompile Include="StringHelpers.cpp" />
@@ -101,7 +100,6 @@
     <ClInclude Include="MemcpyFast.h" />
     <ClInclude Include="Path.h" />
     <ClInclude Include="PrecompiledHeader.h" />
-    <ClInclude Include="pxCheckBox.h" />
     <ClInclude Include="pxEvents.h" />
     <ClInclude Include="pxRadioPanel.h" />
     <ClInclude Include="pxStaticText.h" />
