+++ pcsx2/pcsx2.vcxproj.filters
@@ -1649,6 +1649,9 @@
     <ClCompile Include="gui\CheckedStaticBox.cpp">
       <Filter>AppHost</Filter>
     </ClCompile>
+    <ClCompile Include="gui\pxCheckBox.cpp">
+      <Filter>AppHost</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Patch.h">
@@ -2729,6 +2732,9 @@
       <Filter>AppHost</Filter>
     </ClInclude>
     <ClInclude Include="gui\CheckedStaticBox.h">
+      <Filter>AppHost</Filter>
+    </ClInclude>
+    <ClInclude Include="gui\pxCheckBox.h">
       <Filter>AppHost</Filter>
     </ClInclude>
   </ItemGroup>
